refactor(ledger/state): unify the native amount-round canonicalize (#926)
The muldiv-round consolidation from #926 landed earlier; the one remaining
acceptance item was "one canonicalize pair". The native (XRP-drops) canonicalize
was still duplicated: state.canonicalizeRoundNative vs payment's CanonicalizeDrops
/ CanonicalizeDropsStrict (used by both the offer native path and payment/quality.go).

Promote a single native-round canonicalize into the state package
(CanonicalizeDrops, CanonicalizeDropsStrict, NativeRoundDrops) and have offer and
payment delegate to it; delete payment's duplicate pair. state, the lowest layer,
is the only legal home (payment->state and offer->state both import cleanly).

Behaviour-preserving on every reachable input: the offer and payment native paths
are byte-identical (logic moved, not changed), and MulRoundNative/DivRoundNative
change only on operand pairs whose product exceeds 10^17 drops, where they now
scale a positive residual offset like offer/payment/rippled instead of silently
dropping it. Such amounts cannot occur on chain. The affected native golden rows
were re-captured; IOU columns are unchanged. Conformance: 0 regressions vs the
merge base (204 failing subtests, identical sets).

Fixes #926
