+++ internal/ledger/state/amount_round_golden_test.go
@@ -99,24 +99,29 @@ func TestAmountRoundGolden(t *testing.T) {
 	}
 }
 
-// amountRoundGoldens returns the captured byte-exact outputs of the
-// pre-consolidation implementation, one signature string per operand pair.
+// amountRoundGoldens returns the captured byte-exact outputs, one signature
+// string per operand pair. The IOU columns (MS/M/D/DS) match the pre-#894
+// implementation exactly. The native columns (MN/DN) match it on every input
+// reachable in a valid ledger; they differ only on operand pairs whose product
+// would exceed 10^17 drops, where MulRoundNative/DivRoundNative now scale a
+// positive residual offset like the offer and payment native paths (and like
+// rippled) instead of silently dropping it. Such amounts cannot occur on chain.
 func amountRoundGoldens() []string {
 	return []string{
-		"|ru=false MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3000000000000000,e=-14,neg=false,zero=false) DS=IOU(m=3000000000000000,e=-14,neg=false,zero=false) DN=30 MN=-3520120672398401536|ru=true MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3000000000000001,e=-14,neg=false,zero=false) DS=IOU(m=3000000000000001,e=-14,neg=false,zero=false) DN=30 MN=33333333333333330",
-		"|ru=false MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DS=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DN=0 MN=-3520120672398401536|ru=true MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DS=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DN=1 MN=33333333333333330",
-		"|ru=false MS=IOU(m=7000000000000000,e=20,neg=false,zero=false) M=IOU(m=7000000000000000,e=20,neg=false,zero=false) D=IOU(m=1428571428571428,e=-21,neg=false,zero=false) DS=IOU(m=1428571428571428,e=-21,neg=false,zero=false) DN=0 MN=3509693170864619520|ru=true MS=IOU(m=7000000000000000,e=20,neg=false,zero=false) M=IOU(m=7000000000000000,e=20,neg=false,zero=false) D=IOU(m=1428571428571429,e=-21,neg=false,zero=false) DS=IOU(m=1428571428571429,e=-21,neg=false,zero=false) DN=1 MN=70000000000000000",
-		"|ru=false MS=IOU(m=-1750000000000000,e=11,neg=true,zero=false) M=IOU(m=-1750000000000000,e=11,neg=true,zero=false) D=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DS=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DN=-2800000000000001 MN=-175000000000000000|ru=true MS=IOU(m=-1750000000000000,e=11,neg=true,zero=false) M=IOU(m=-1750000000000000,e=11,neg=true,zero=false) D=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DS=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DN=-2800000000000000 MN=-170598510725431296",
-		"|ru=false MS=IOU(m=-8333333333333333,e=3,neg=true,zero=false) M=IOU(m=-8333333333333333,e=3,neg=true,zero=false) D=IOU(m=-7500000000000001,e=-24,neg=true,zero=false) DS=IOU(m=-7500000000000001,e=-24,neg=true,zero=false) DN=0 MN=-83333333333333325|ru=true MS=IOU(m=-8333333333333332,e=3,neg=true,zero=false) M=IOU(m=-8333333333333332,e=3,neg=true,zero=false) D=IOU(m=-7500000000000000,e=-24,neg=true,zero=false) DS=IOU(m=-7500000000000000,e=-24,neg=true,zero=false) DN=0 MN=-8333333333333332500",
-		"|ru=false MS=PANIC(IOUAmount overflow) M=PANIC(IOUAmount overflow) D=IOU(m=9999999999999999,e=65,neg=false,zero=false) DS=IOU(m=9999999999999999,e=65,neg=false,zero=false) DN=0 MN=0|ru=true MS=PANIC(IOUAmount overflow) M=PANIC(IOUAmount overflow) D=IOU(m=9999999999999999,e=65,neg=false,zero=false) DS=IOU(m=9999999999999999,e=65,neg=false,zero=false) DN=999999999999999900 MN=99999999999999990",
+		"|ru=false MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3000000000000000,e=-14,neg=false,zero=false) DS=IOU(m=3000000000000000,e=-14,neg=false,zero=false) DN=30 MN=-3520120672398401536|ru=true MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3000000000000001,e=-14,neg=false,zero=false) DS=IOU(m=3000000000000001,e=-14,neg=false,zero=false) DN=30 MN=-3520120672398401536",
+		"|ru=false MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DS=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DN=0 MN=-3520120672398401536|ru=true MS=IOU(m=3333333333333333,e=13,neg=false,zero=false) M=IOU(m=3333333333333333,e=13,neg=false,zero=false) D=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DS=IOU(m=3333333333333333,e=-17,neg=false,zero=false) DN=1 MN=-3520120672398401536",
+		"|ru=false MS=IOU(m=7000000000000000,e=20,neg=false,zero=false) M=IOU(m=7000000000000000,e=20,neg=false,zero=false) D=IOU(m=1428571428571428,e=-21,neg=false,zero=false) DS=IOU(m=1428571428571428,e=-21,neg=false,zero=false) DN=0 MN=3509693170864619520|ru=true MS=IOU(m=7000000000000000,e=20,neg=false,zero=false) M=IOU(m=7000000000000000,e=20,neg=false,zero=false) D=IOU(m=1428571428571429,e=-21,neg=false,zero=false) DS=IOU(m=1428571428571429,e=-21,neg=false,zero=false) DN=1 MN=3509693170864619520",
+		"|ru=false MS=IOU(m=-1750000000000000,e=11,neg=true,zero=false) M=IOU(m=-1750000000000000,e=11,neg=true,zero=false) D=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DS=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DN=-2800000000000001 MN=-170598510725431296|ru=true MS=IOU(m=-1750000000000000,e=11,neg=true,zero=false) M=IOU(m=-1750000000000000,e=11,neg=true,zero=false) D=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DS=IOU(m=-2800000000000000,e=0,neg=true,zero=false) DN=-2800000000000000 MN=-170598510725431296",
+		"|ru=false MS=IOU(m=-8333333333333333,e=3,neg=true,zero=false) M=IOU(m=-8333333333333333,e=3,neg=true,zero=false) D=IOU(m=-7500000000000001,e=-24,neg=true,zero=false) DS=IOU(m=-7500000000000001,e=-24,neg=true,zero=false) DN=0 MN=-8333333333333332500|ru=true MS=IOU(m=-8333333333333332,e=3,neg=true,zero=false) M=IOU(m=-8333333333333332,e=3,neg=true,zero=false) D=IOU(m=-7500000000000000,e=-24,neg=true,zero=false) DS=IOU(m=-7500000000000000,e=-24,neg=true,zero=false) DN=0 MN=-8333333333333332500",
+		"|ru=false MS=PANIC(IOUAmount overflow) M=PANIC(IOUAmount overflow) D=IOU(m=9999999999999999,e=65,neg=false,zero=false) DS=IOU(m=9999999999999999,e=65,neg=false,zero=false) DN=0 MN=0|ru=true MS=PANIC(IOUAmount overflow) M=PANIC(IOUAmount overflow) D=IOU(m=9999999999999999,e=65,neg=false,zero=false) DS=IOU(m=9999999999999999,e=65,neg=false,zero=false) DN=1 MN=1",
 		"|ru=false MS=IOU(m=1000000000000001,e=-81,neg=false,zero=false) M=IOU(m=1000000000000001,e=-81,neg=false,zero=false) D=IOU(m=0,e=-100,neg=false,zero=true) DS=IOU(m=0,e=-100,neg=false,zero=true) DN=0 MN=0|ru=true MS=IOU(m=1000000000000001,e=-81,neg=false,zero=false) M=IOU(m=1000000000000001,e=-81,neg=false,zero=false) D=IOU(m=1000000000000000,e=-96,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-96,neg=false,zero=false) DN=1 MN=1",
 		"|ru=false MS=IOU(m=7000000000000000,e=-9,neg=false,zero=false) M=IOU(m=7000000000000000,e=-9,neg=false,zero=false) D=IOU(m=1428571428571428,e=-10,neg=false,zero=false) DS=IOU(m=1428571428571428,e=-10,neg=false,zero=false) DN=142857 MN=7000000|ru=true MS=IOU(m=7000000000000000,e=-9,neg=false,zero=false) M=IOU(m=7000000000000000,e=-9,neg=false,zero=false) D=IOU(m=1428571428571429,e=-10,neg=false,zero=false) DS=IOU(m=1428571428571429,e=-10,neg=false,zero=false) DN=142858 MN=7000000",
 		"|ru=false MS=IOU(m=7000000000000000,e=-9,neg=false,zero=false) M=IOU(m=7000000000000000,e=-9,neg=false,zero=false) D=IOU(m=7000000000000000,e=-21,neg=false,zero=false) DS=IOU(m=7000000000000000,e=-21,neg=false,zero=false) DN=0 MN=7000000|ru=true MS=IOU(m=7000000000000000,e=-9,neg=false,zero=false) M=IOU(m=7000000000000000,e=-9,neg=false,zero=false) D=IOU(m=7000000000000000,e=-21,neg=false,zero=false) DS=IOU(m=7000000000000000,e=-21,neg=false,zero=false) DN=1 MN=7000000",
 		"|ru=false MS=IOU(m=-1234567890000000,e=-1,neg=true,zero=false) M=IOU(m=-1234567890000000,e=-1,neg=true,zero=false) D=IOU(m=-1234567890000000,e=-13,neg=true,zero=false) DS=IOU(m=-1234567890000000,e=-13,neg=true,zero=false) DN=-124 MN=-123456789000001|ru=true MS=IOU(m=-1234567890000000,e=-1,neg=true,zero=false) M=IOU(m=-1234567890000000,e=-1,neg=true,zero=false) D=IOU(m=-1234567890000000,e=-13,neg=true,zero=false) DS=IOU(m=-1234567890000000,e=-13,neg=true,zero=false) DN=-123 MN=-123456789000000",
 		"|ru=false MS=IOU(m=7000000000000000,e=-4,neg=false,zero=false) M=IOU(m=7000000000000000,e=-4,neg=false,zero=false) D=IOU(m=1428571428571428,e=-5,neg=false,zero=false) DS=IOU(m=1428571428571428,e=-5,neg=false,zero=false) DN=14285714285 MN=700000000000|ru=true MS=IOU(m=7000000000000000,e=-4,neg=false,zero=false) M=IOU(m=7000000000000000,e=-4,neg=false,zero=false) D=IOU(m=1428571428571429,e=-5,neg=false,zero=false) DS=IOU(m=1428571428571429,e=-5,neg=false,zero=false) DN=14285714286 MN=700000000000",
-		"|ru=false MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DN=1000000000 MN=3875820019684212736|ru=true MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DN=1000000000 MN=10000000000000000",
-		"|ru=false MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DN=0 MN=3875820019684212736|ru=true MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DN=1 MN=10000000000000000",
+		"|ru=false MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DN=1000000000 MN=3875820019684212736|ru=true MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-6,neg=false,zero=false) DN=1000000000 MN=3875820019684212736",
+		"|ru=false MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DN=0 MN=3875820019684212736|ru=true MS=IOU(m=1000000000000000,e=6,neg=false,zero=false) M=IOU(m=1000000000000000,e=6,neg=false,zero=false) D=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DS=IOU(m=1000000000000000,e=-24,neg=false,zero=false) DN=1 MN=3875820019684212736",
 		"|ru=false MS=IOU(m=3086419725000000,e=-2,neg=false,zero=false) M=IOU(m=3086419725000000,e=-2,neg=false,zero=false) D=IOU(m=2025000018427500,e=-18,neg=false,zero=false) DS=IOU(m=2025000018427500,e=-18,neg=false,zero=false) DN=0 MN=30864197250000|ru=true MS=IOU(m=3086419725000000,e=-2,neg=false,zero=false) M=IOU(m=3086419725000000,e=-2,neg=false,zero=false) D=IOU(m=2025000018427501,e=-18,neg=false,zero=false) DS=IOU(m=2025000018427501,e=-18,neg=false,zero=false) DN=1 MN=30864197250000",
-		"|ru=false MS=IOU(m=7000000000000000,e=16,neg=false,zero=false) M=IOU(m=7000000000000000,e=16,neg=false,zero=false) D=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DN=7000000000 MN=4897961520886972416|ru=true MS=IOU(m=7000000000000000,e=16,neg=false,zero=false) M=IOU(m=7000000000000000,e=16,neg=false,zero=false) D=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DN=7000000000 MN=70000000000000000",
+		"|ru=false MS=IOU(m=7000000000000000,e=16,neg=false,zero=false) M=IOU(m=7000000000000000,e=16,neg=false,zero=false) D=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DN=7000000000 MN=4897961520886972416|ru=true MS=IOU(m=7000000000000000,e=16,neg=false,zero=false) M=IOU(m=7000000000000000,e=16,neg=false,zero=false) D=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DS=IOU(m=7000000000000000,e=-6,neg=false,zero=false) DN=7000000000 MN=4897961520886972416",
 	}
 }
